repos/Linux Compliance Report Automation/cis-compliance-checker/requirements.txt
python-nmap==0.7.1
beautifulsoup4==4.12.2
lxml==4.9.3
